[test] XML parser moved to separate file
